requirements.txt
pytest==3.0.7
Flask==1.0.2
pylint==1.6.4
gunicorn==19.9.0
pytest-cov==2.5.1
coveralls==1.3.0
nose==1.3.7
flask-restful==0.3.6
flask-bcrypt==0.7.1
pyjwt==1.6.4
testing.postgresql==1.3.0
